+++ api/src/db/seeds/test/users.js
@@ -1,12 +1,13 @@
 const usersClock = require('../../../users_clock')
-const { OSMESA_DB } = require('../../../config')
-const osmesaDB = require('knex')({
-  client: 'pg',
-  connection: OSMESA_DB
-})
+const { cache } = require('../../../config')
 
 exports.seed = (knex) => knex('users').del() // delete entries
   .then(async () => {
+    const osmesaDB = require('knex')({
+      client: 'pg',
+      connection: cache.get('osmesa-db')
+    })
+
     const users = await osmesaDB('users').select('id as osm_id', 'name as full_name')
     return knex('users').insert(users)
   })