+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Memory cache for settings
 - Added blank state illustration for users with no edits
 - Added "Total" row for campaign tables, calculating all campaign stat totals
+- Add CLI tool to allow user to assign random edits to their user for development
 
 ### Changed 
 - User extent tiles now provided by s3 and tileserver proxy